requirements.txt
streamlit 
flask 
pandas
google-auth
google-auth-oauthlib 
google-auth-httplib2
google-api-python-client
openai 
serpapi
requests